+++ ccsdspy/constants.py
@@ -1,0 +1,7 @@
+"""Constants used in the CCSDSPy package."""
+
+# Number of bits in a byte
+BITS_PER_BYTE = 8
+
+# Number of bytes in a CCSDS primary headero
+PRIMARY_HEADER_NUM_BYTES = 6